server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain-core==0.3.46
langchain-openai==0.3.9
langsmith==0.3.18
MarkupSafe==3.0.2
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.3
openai==1.67.0
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.15.2
seaborn==0.13.2
setuptools==77.0.3
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
supervision==0.25.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.95
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
zstandard==0.23.0
